requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
coverage==7.13.1
cron-descriptor==2.0.6
dj-database-url==3.1.0
Django==5.2.10
django-celery-beat==2.8.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-filter==25.2
django-redis==6.0.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.11
joblib==1.5.3
kombu==5.6.2
numpy==2.4.1
packaging==25.0
pandas==2.3.3
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
